+++ Item.h
@@ -3,16 +3,16 @@
 class CItem
 {
 public:
-	CItem() = delete;
+	CItem()  = delete;
 	~CItem() = default;
 
-	CItem(char f_sku, int f_price) : m_sku(f_sku), m_price(f_price) {}
+	CItem(char f_sku, float f_price) : m_sku(f_sku), m_price(f_price) {}
 
-	char getSKU();
-	int  getPrice();
+	char   getSKU();
+	float  getPrice();
 
 private:
-	const char m_sku;
-	const int  m_price;
+	const char   m_sku;
+	const float  m_price;
 };
